requirements.txt
Flask==3.0.3
mysql-connector-python==8.4.0
requests==2.31.0
tmdbv3api==1.9.0
